Reset
After tinkering for days, I've decided against a wholesale redesign for
a more iterative approach.

These are the theme files for my site as it exists today. I'm hoping to
add support for a portfolio, image posts (a la Tumblr), and maybe a
"big footer" in the coming days & weeks. We'll see how that goes.